+++ CHANGELOG.rst
@@ -3,7 +3,6 @@ Change Log
 
 [TODO]
 --------------------
-- [???] example (and test) on how to corrupt the observation for the agent (without corrupting the environment)
 - [???] use some kind of "env.get_state()" when simulating instead of recoding everything "by hand"
 - [???] use "backend.get_action_to_set()" in simulate
 - [???] use the prod_p_forecasted and co in the "next_chronics" of simulate
@@ -32,6 +31,12 @@ Change Log
 
 [1.6.5] - 2022-xx-yy
 ---------------------
+- [BREAKING] the function "env.reset()" now reset the underlying pseudo random number generators
+  of all the environment subclasses (eg. observation space, action space, etc.) This change has been made to
+  ensure reproducibility between episodes: if `env.seed(...)` is called once, then regardless of what happens
+  (basically the number of "env.step()" between calls to "env.reset()")
+  the "env.reset()" will be generated with the same prng (drawn from the environment)
+  This effect the opponent and the chronics (when maintenance are generated "on the fly").
 - [BREAKING] the name of the python files for the "Chronics" module are now lowercase (complient with PEP). If you
   did things like `from grid2op.Chronics.ChangeNothing import ChangeNothing` you need to change it like
   `from grid2op.Chronics.changeNothing import ChangeNothing` or even better, and this is the preferred way to include
@@ -39,6 +44,8 @@ Change Log
   are to be expected in following versions).
 - [BREAKING] same as above for the "Observation" module. It should not affect lots of code (more refactoring of the kind
   are to be expected in following versions).
+- [FIXED] a bug for the EpisodeData that did not save the first observation when 
+  "add_detailed_output" was set to ``True`` and the data were not saved on disk.
 - [FIXED] an issue when copying the environment with the opponent (see issue https://github.com/rte-france/Grid2Op/issues/274)
 - [FIXED] a bug leading to the wrong "backend.get_action_to_set()" when there were storage units on the grid. 
 - [FIXED] a bug in the "BackendConverter" when there are storage  on the grid
@@ -56,6 +63,9 @@ Change Log
   `env.action_space.get_back_to_ref_state(obs)`
 - [ADDED] a method of the action to store it in a grid2op independant fashion (using json and dictionaries), see `act.as_serializable_dict()`
 - [ADDED] possibility to generate a gym `DiscreteActSpace` from a given list of actions (see https://github.com/rte-france/Grid2Op/issues/277)
+- [ADDED] a class that output a noisy observation to the agent (see `NoisyObservation`): the agent sees
+  the real values of the environment with some noise, this could used to model inacurate
+  sensors.
 - [IMPROVED] observation now raises `Grid2OpException` instead of `RuntimeError`
 - [IMRPOVED] docs (and notebooks) for the "split_train_val" https://github.com/rte-france/Grid2Op/issues/269
 - [IMRPOVED] the "split_train_val" function to also generate a test dataset see https://github.com/rte-france/Grid2Op/issues/276